move common and optional fields up to ServerInfo
